refactor: Reorganize code dir
