+++ chidian/data_mapping.py
@@ -6,6 +6,12 @@ from typing import Any, Generic, Type, TypeVar
 
 from pydantic import BaseModel
 
+from .lib.data_mapping_helpers import (
+    to_dict,
+    validate_input,
+    validate_output,
+    validate_schemas,
+)
 from .mapper import Mapper
 
 # Define generic type variables bounded to BaseModel
@@ -37,7 +43,7 @@ class DataMapping(Generic[_InModel, _OutModel]):
             output_schema: Pydantic BaseModel class for output validation
             strict: If True, enforce strict validation
         """
-        self._validate_schemas(input_schema, output_schema)
+        validate_schemas(input_schema, output_schema)
 
         self.mapper = mapper
         self.input_schema = input_schema
@@ -58,55 +64,13 @@ class DataMapping(Generic[_InModel, _OutModel]):
             ValidationError: If input or output validation fails
         """
         # Validate and convert input
-        validated_input = self._validate_input(data)
+        validated_input = validate_input(data, self.input_schema)
 
         # Convert to dict for Mapper
-        input_dict = self._to_dict(validated_input)
+        input_dict = to_dict(validated_input)
 
         # Apply transformation
         output_dict = self.mapper(input_dict)
 
         # Validate and return output
-        return self._validate_output(output_dict)
-
-    def _validate_schemas(self, input_schema: Type, output_schema: Type) -> None:
-        """Validate that schemas are Pydantic BaseModel classes."""
-        if not self._is_pydantic_model(input_schema):
-            raise TypeError(
-                f"input_schema must be a Pydantic BaseModel, got {type(input_schema)}"
-            )
-        if not self._is_pydantic_model(output_schema):
-            raise TypeError(
-                f"output_schema must be a Pydantic BaseModel, got {type(output_schema)}"
-            )
-
-    def _is_pydantic_model(self, model_class: Type) -> bool:
-        """Check if a class is a Pydantic BaseModel."""
-        try:
-            return (
-                isinstance(model_class, type)
-                and issubclass(model_class, BaseModel)
-                and hasattr(model_class, "model_fields")
-            )
-        except TypeError:
-            return False
-
-    def _validate_input(self, data: Any) -> _InModel:
-        """Validate input data against input schema."""
-        if isinstance(data, self.input_schema):
-            return data  # type: ignore[return-value]
-
-        # Try to convert dict to model
-        if isinstance(data, dict):
-            return self.input_schema.model_validate(data)  # type: ignore[return-value]
-
-        # Try direct validation
-        return self.input_schema.model_validate(data)  # type: ignore[return-value]
-
-    def _to_dict(self, model: _InModel) -> dict[str, Any]:
-        """Convert Pydantic model to dictionary."""
-        return model.model_dump()
-
-    def _validate_output(self, data: dict[str, Any]) -> _OutModel:
-        """Validate output data against output schema."""
-        return self.output_schema.model_validate(data)  # type: ignore[return-value]
+        return validate_output(output_dict, self.output_schema)
