+++ chidian/lib/__init__.py
@@ -1,0 +1,5 @@
+"""
+Internal library modules for chidian.
+
+This package contains helper functions, parsers, and implementation details.
+"""
